requirements.txt
adafruit-circuitpython-max31865
adafruit-circuitpython-ads1x15
adafruit-circuitpython-busdevice
h5py
RPI.GPIO
matplotlib
pyserial
numpy
filelock
